Extract folder browser drag drop
